+++ src/LEGO.AsyncAPI.Bindings/ChannelBinding.cs
@@ -11,39 +11,27 @@ namespace LEGO.AsyncAPI.Bindings
     {
         protected abstract FixedFieldMap<T> FixedFieldMap { get; }
 
-        public override IServerBinding Parse(PropertyNode parseNode)
-        {
-            return LoadBinding("ServerBinding", parseNode, FixedFieldMap);
-        }
+        public override IServerBinding LoadBinding(PropertyNode parseNode) => LoadBinding("ServerBinding", parseNode, FixedFieldMap);
     }
 
     public abstract class ChannelBinding<T> : BindingDeserializer, IBindingParser<IChannelBinding> where T : IChannelBinding, new()
     {
         protected abstract FixedFieldMap<T> FixedFieldMap { get; }
 
-        public override IChannelBinding Parse(PropertyNode node)
-        {
-            return LoadBinding("ChannelBinding", node.Value, FixedFieldMap);
-        }
+        public override IChannelBinding LoadBinding(PropertyNode node) => LoadBinding("ChannelBinding", node.Value, FixedFieldMap);
     }
 
     public abstract class OperationBinding<T> : BindingDeserializer, IBindingParser<IOperationBinding> where T : IOperationBinding, new()
     {
         protected abstract FixedFieldMap<T> FixedFieldMap { get; }
 
-        public override IOperationBinding Parse(PropertyNode parseNode)
-        {
-            return LoadBinding("OperationBinding", parseNode, FixedFieldMap);
-        }
+        public override IOperationBinding LoadBinding(PropertyNode parseNode) => LoadBinding("OperationBinding", parseNode, FixedFieldMap);
     }
 
     public abstract class MessageBinding<T> : BindingDeserializer, IBindingParser<IMessageBinding> where T : IMessageBinding, new()
     {
         protected abstract FixedFieldMap<T> FixedFieldMap { get; }
 
-        public override IMessageBinding Parse(PropertyNode parseNode)
-        {
-            return LoadBinding("MessageBinding", parseNode, FixedFieldMap);
-        }
+        public override IMessageBinding LoadBinding(PropertyNode parseNode) => LoadBinding("MessageBinding", parseNode, FixedFieldMap);
     }
 }
